backend-django/requirements.txt
asgiref==3.2.3
Django==2.2.7
django-cors-headers==3.2.1
django-default-imagefield==0.5.5
django-polymorphic==2.1.2
djangorestframework==3.11.0
npm==0.1.1
optional-django==0.1.0
Pillow==6.2.2
pytz==2019.3
sqlparse==0.3.0
